Index page translated
Added translation for Turkish

+++ database/migrations/2015_04_19_093814_create_reviews_table.php
@@ -21,7 +21,7 @@ class CreateReviewsTable extends Migration {
 			$table->timestamp('date')->default(DB::raw('CURRENT_TIMESTAMP'));
 			$table->foreign('fromUser')->references('id')->on('users');
 			$table->foreign('toUser')->references('id')->on('users');
-			$table->integer('grade')->nullable();
+			$table->double('grade')->nullable();
 		});
 	}
